+++ CHANGELOG.md
@@ -24,6 +24,7 @@ v0.15.19 (??.??.????)
  * new Wrapper function tx_rnbase_mod_Util::mapRecordTypeToSpriteIconName
  * new Wrapper function tx_rnbase_parameters::setGetParameter
  * new Wrapper function tx_rnbase_mod_BaseModule::issueCommand
+ * new class tx_rnbase_util_Typo3Classes to get TYPO3 classes independent of TYPO3 version
 
 v0.15.18 (02.12.2015)
  * Bugfix: tx_rnbase_util_Dates::date_mysql2tstamp liefert nun immer einen Integer und nicht NULL im Fehlerfall